HasNavigationHeadings trait
