+++ app/Models/Article.php
@@ -3,12 +3,11 @@
 namespace App\Models;
 
 use App\Traits\Publishable;
+use App\Traits\HasNavigationHeadings;
 use Backpack\CRUD\app\Models\Traits\CrudTrait;
 use Backpack\CRUD\app\Models\Traits\SpatieTranslatable\HasTranslations;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\UploadedFile;
-use Illuminate\Support\Str;
-use PHPHtmlParser\Dom;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 use Spatie\Sluggable\HasSlug;
@@ -16,23 +15,11 @@ use Spatie\Sluggable\SlugOptions;
 
 class Article extends Model implements HasMedia
 {
-    use CrudTrait, InteractsWithMedia, HasSlug, HasTranslations, Publishable;
+    use CrudTrait, InteractsWithMedia, HasSlug, HasTranslations, Publishable, HasNavigationHeadings;
 
     protected $guarded = ['id'];
     protected $dates = ['published_at'];
     protected $translatable = ['title', 'content'];
-
-    public function getContentHtmlAttribute()
-    {
-        $dom = new Dom;
-        $dom->loadStr($this->content);
-
-        foreach($this->findHeadingsForNavigation($dom) as $index => $heading) {
-            $heading->setAttribute('id', Str::slug($heading->text) . "-$index");
-        }
-
-        return $dom;
-    }
 
     public function getCoverImageTagAttribute()
     {
@@ -54,18 +41,6 @@ class Article extends Model implements HasMedia
             );
     }
 
-    public function getNavigationHeadingsAttribute()
-    {
-        $dom = new Dom;
-        $dom->loadStr($this->content);
-
-        return collect($this->findHeadingsForNavigation($dom))
-            ->map(fn ($heading, $index) => (object) [
-                'text' => $heading->text,
-                'href' => '#' . Str::slug($heading->text) . "-$index",
-            ]);
-    }
-
     public function getRouteKeyName()
     {
         return 'slug';
@@ -83,10 +58,5 @@ class Article extends Model implements HasMedia
         $this
             ->addMediaCollection('default')
             ->withResponsiveImages();
-    }
-
-    private function findHeadingsForNavigation($dom)
-    {
-        return $dom->find('h1,h2,h3');
     }
 }
